Supplier referer can modify part of another gasmember's basket

+++ gasistafelice/gas/forms/order/gmo.py
@@ -41,21 +41,22 @@ class BaseGASMemberOrderForm(forms.Form):
             )
 
         if self._gmusr != self._loggedusr:
-            log.debug(u"---- %s (%s) not enabled for %s" % (
-                self.__class__.__name__,
-                self._gmusr, self._loggedusr
-            ))
             #DELEGATE: order.referrer_person can make order in name of other person
             #In this case we can authorize and set in the note the person who act the gasmemberorder
             id = self.cleaned_data.get('id')
             if id:
                 gmo = GASMemberOrder.objects.get(pk=id)
-                if not self._loggedusr == gmo.order.referrer_person.user:
-                    raise forms.ValidationError(
-                        _(u"You %(logged)s are not authorized to make an order for %(person)s") % {
-                            'logged' : u"(%s)" % self._loggedusr, 
-                            'person' :self._gmusr
-                    })
+                if gmo.ordered_amount != self.cleaned_data.get('ordered_amount'): 
+                    if not self._loggedusr == gmo.order.referrer_person.user:
+                        log.debug(u"---- %s (%s) BASE VALIDATION not enabled for %s" % (
+                            self.__class__.__name__,
+                            self._gmusr, self._loggedusr
+                        ))
+                        raise forms.ValidationError(
+                            _(u"You %(logged)s are not authorized to make an order for %(person)s") % {
+                                'logged' : u"(%s)" % self._loggedusr, 
+                                'person' :self._gmusr
+                        })
         return cleaned_data
 
 
@@ -83,6 +84,10 @@ class SingleGASMemberOrderForm(BaseGASMemberOrderForm):
                 gmo.note = self.cleaned_data.get('note')
 
                 if self._gmusr != self._loggedusr and self._loggedusr == gmo.order.referrer_person.user:
+                    log.debug(u"---- %s (%s) DELEGATE UPDATE for %s" % (
+                        self.__class__.__name__,
+                        self._gmusr, self._loggedusr
+                    ))
                     delegate = _("[ord by %s] ") % gmo.order.referrer_person.report_name
                     if gmo.note.find(delegate) == -1:
                         gmo.note = delegate + gmo.note
@@ -99,6 +104,10 @@ class SingleGASMemberOrderForm(BaseGASMemberOrderForm):
                     gsop = GASSupplierOrderProduct.objects.get(pk=self.cleaned_data.get('gsop_id'))
                     note = self.cleaned_data.get('note')
                     if self._gmusr != self._loggedusr and self._loggedusr == gsop.order.referrer_person.user:
+                        log.debug(u"---- %s (%s) DELEGATE CREATE for %s" % (
+                            self.__class__.__name__,
+                            self._gmusr, self._loggedusr
+                        ))
                         delegate = _("[ord by %s] ") % gsop.order.referrer_person.report_name
                         if note.find(delegate) == -1:
                             note = delegate + note
@@ -145,21 +154,46 @@ class BasketGASMemberOrderForm(BaseGASMemberOrderForm):
     #            if gm_id and gm_id != gmo.purchaser.pk:
     #                print "Qualcosa non va con: GASmember"
     #                return ""
-                gmo.ordered_price = self.cleaned_data.get('ordered_price')
-                gmo.ordered_amount = self.cleaned_data.get('ordered_amount')
-                # log.debug("BasketGASMemberOrderForm (%s) enabled = %s" % (gmo.pk,enabled))
-                if gmo.ordered_amount == 0:
-                    log.debug(u"REMOVING GASMemberOrder (%s) from BASKET using amount widget (+ -)")
-                    gmo.delete()
-                    log.debug(u"REMOVED")
-                elif enabled:
-                    log.debug(u"REMOVING GASMemberOrder (%s) from BASKET using check enabled")
-                    gmo.delete()
-                    log.debug(u"REMOVED")
-                else:
-                    log.debug(u"UPDATING GASMemberOrder (%s) from BASKET" % id)
-                    gmo.save()
-                    log.debug(u"UPDATED")
+                #On basket do nothing if no change needed
+                if gmo.ordered_amount != self.cleaned_data.get('ordered_amount') or enabled: 
+
+                    if not self._loggedusr == gmo.order.referrer_person.user:
+                        log.debug(u"---- %s (%s) BASKET EXECUTING not enabled for %s" % (
+                            self.__class__.__name__,
+                            self._gmusr, self._loggedusr
+                        ))
+                        raise forms.ValidationError(
+                            _(u"You %(logged)s are not authorized to make an order for %(person)s") % {
+                                'logged' : u"(%s)" % self._loggedusr, 
+                                'person' :self._gmusr
+                        })
+                        return;
+                    else:
+                        delegate = _("[ord by %s] ") % gmo.order.referrer_person.report_name
+                        if gmo.note.find(delegate) == -1:
+                            gmo.note = delegate + gmo.note
+                        log.debug(u"---- %s (%s) DELEGATE change BASKET for %s" % (
+                            self.__class__.__name__,
+                            self._gmusr, self._loggedusr
+                        ))
+                
+                    gmo.ordered_price = self.cleaned_data.get('ordered_price')
+                    gmo.ordered_amount = self.cleaned_data.get('ordered_amount')
+                    # log.debug("BasketGASMemberOrderForm (%s) enabled = %s" % (gmo.pk,enabled))
+                    if gmo.ordered_amount == 0:
+                        log.debug(u"REMOVING GASMemberOrder (%s) from BASKET using amount widget (+ -)")
+                        gmo.delete()
+                        log.debug(u"REMOVED")
+                    elif enabled:
+                        log.debug(u"REMOVING GASMemberOrder (%s) from BASKET using check enabled")
+                        gmo.delete()
+                        log.debug(u"REMOVED")
+                    else:
+                        log.debug(u"UPDATING GASMemberOrder (%s) from BASKET" % id)
+                        gmo.save()
+                        log.debug(u"UPDATED")
+                #else:
+                #    log.debug(u"NOTHING TODO NOTHING")
         else:
             log.warning("BasketGASMemberOrderForm.save(): form is not valid. is_valid() SHOULD be called before calling save()")
